+++ crates/forge/tests/integration/mod.rs
@@ -2,7 +2,7 @@ pub(crate) mod common;
 mod declare;
 mod deploy;
 mod dispatchers;
-mod emit_events;
+mod expect_events;
 mod prank;
 mod pure_cairo;
 mod roll;
